A function that reads a text file and prints it to the POSIZ standard output

+++ 0x15-file_io/0-read_textfile.c
@@ -9,7 +9,7 @@
 */
 ssize_t read_textfile(const char *filename, size_t letters)
 {
-        size_t toOpen, toRead, toWrite;
+        int toOpen, toRead, toWrite;
         char *buf;
 
         /**
